fix(charts): centralize chart browser commands (#264)
