+++ vitest.config.ts
@@ -1,6 +1,5 @@
 import { playwright } from "@vitest/browser-playwright";
-import axeCore from "axe-core";
-import type { Frame, FrameLocator, Locator } from "playwright";
+import type { Frame } from "playwright";
 import solid from "vite-plugin-solid";
 import { defineConfig } from "vitest/config";
 import { chartBrowserCommands } from "./tests/browser/support/chart-commands";
@@ -25,276 +24,6 @@ async function getProductRouteTestFrame(context: unknown) {
   const header = routeFrame.locator("[data-product-header]");
   await header.waitFor({ state: "visible" });
   return { testFrame, routeFrame, header };
-}
-
-type ChartCatalogFixture = {
-  routeTitle: string;
-  firstSlug: string;
-  firstExportName: string;
-  interactiveSlug: string;
-  renderedSeriesSelector: string;
-  recoveredPreviewTitle: string;
-};
-
-const chartCatalogFixtures = {
-  area: {
-    routeTitle: "Chart Catalog route",
-    firstSlug: "chart-area-axes",
-    firstExportName: "ChartAreaAxes",
-    interactiveSlug: "chart-area-interactive",
-    renderedSeriesSelector: ".recharts-area-area",
-    recoveredPreviewTitle: "Area Chart — Axes Preview",
-  },
-  line: {
-    routeTitle: "Line Chart Catalog route",
-    firstSlug: "chart-line-default",
-    firstExportName: "ChartLineDefault",
-    interactiveSlug: "chart-line-interactive",
-    renderedSeriesSelector: ".recharts-line-curve",
-    recoveredPreviewTitle: "Line Chart Preview",
-  },
-  radar: {
-    routeTitle: "Radar Chart Catalog route",
-    firstSlug: "chart-radar-default",
-    firstExportName: "ChartRadarDefault",
-    interactiveSlug: "chart-radar-default",
-    renderedSeriesSelector: ".recharts-polygon",
-    recoveredPreviewTitle: "Radar Chart Preview",
-  },
-} as const satisfies Record<"area" | "line" | "radar", ChartCatalogFixture>;
-
-async function getBrowserCommandFrame(context: unknown) {
-  const commandContext = context as {
-    provider: {
-      getCommandsContext: (sessionId: string) => {
-        frame: () => Promise<Frame>;
-      };
-    };
-    sessionId: string;
-  };
-  return commandContext.provider.getCommandsContext(commandContext.sessionId).frame();
-}
-
-type ChartFamilyInspectionContext = {
-  testFrame: Frame;
-  routeFrame: FrameLocator;
-  interactiveEntry: Locator;
-};
-
-async function inspectChartCatalogFamily<T>(
-  context: unknown,
-  fixture: ChartCatalogFixture,
-  inspectFamily: (context: ChartFamilyInspectionContext) => Promise<T>,
-) {
-  const testFrame = await getBrowserCommandFrame(context);
-  const page = testFrame.page();
-  const routeFrame = testFrame.frameLocator(`iframe[title="${fixture.routeTitle}"]`);
-  const consoleErrors: string[] = [];
-  const recordConsole = (message: { type: () => string; text: () => string }) => {
-    if (message.type() === "error") consoleErrors.push(message.text());
-  };
-  const recordPageError = (error: Error) => consoleErrors.push(error.message);
-  page.on("console", recordConsole);
-  page.on("pageerror", recordPageError);
-
-  try {
-    await routeFrame.locator("body").evaluate(() => window.location.reload());
-    await testFrame.waitForTimeout(500);
-
-    const heading = routeFrame.getByRole("heading", { name: "Beautiful Charts & Graphs" });
-    await heading.waitFor({ state: "visible" });
-    const entries = routeFrame.locator("[data-chart-entry]");
-    const firstEntry = entries.first();
-    const previews: Array<{
-      slug: string | null;
-      renderedSeriesCount: number;
-      chartRole: string | null;
-      canonicalPath: string | null;
-    }> = [];
-
-    for (let index = 0; index < (await entries.count()); index += 1) {
-      const catalogEntry = entries.nth(index);
-      await catalogEntry.scrollIntoViewIfNeeded();
-      const preview = catalogEntry.frameLocator("iframe");
-      const chart = preview.locator('[data-slot="chart"]');
-      await chart.waitFor({ state: "visible", timeout: 8_000 });
-      await chart.locator(".recharts-surface").waitFor({ state: "visible" });
-      const canonicalHref = await preview.locator('link[rel="canonical"]').getAttribute("href");
-      previews.push({
-        slug: await catalogEntry.getAttribute("data-chart-entry"),
-        renderedSeriesCount: await chart.locator(fixture.renderedSeriesSelector).count(),
-        chartRole: await preview.getByRole("application").getAttribute("role"),
-        canonicalPath: canonicalHref ? new URL(canonicalHref).pathname : null,
-      });
-    }
-
-    await firstEntry.scrollIntoViewIfNeeded();
-    const previewFrame = firstEntry.frameLocator("iframe");
-    const chart = previewFrame.locator('[data-slot="chart"]');
-    await chart.waitFor({ state: "visible" });
-    const rtlHasNoOverflow = await previewFrame.locator("html").evaluate((element) => {
-      (element as HTMLHtmlElement).dir = "rtl";
-      const hasNoOverflow = element.scrollWidth <= element.clientWidth;
-      (element as HTMLHtmlElement).dir = "";
-      return hasNoOverflow;
-    });
-
-    await previewFrame.getByRole("application").focus();
-    await page.keyboard.press("ArrowRight");
-    await testFrame.waitForTimeout(250);
-    const keyboardTooltipText = await previewFrame.locator(".cn-chart-tooltip").textContent();
-    await chart.locator(".recharts-surface").hover({ position: { x: 220, y: 150 } });
-    await previewFrame.locator(".cn-chart-tooltip").waitFor({ state: "visible" });
-    const pointerTooltipText = await previewFrame.locator(".cn-chart-tooltip").textContent();
-
-    const interactiveEntry = routeFrame.locator(`[data-chart-entry="${fixture.interactiveSlug}"]`);
-    await interactiveEntry.scrollIntoViewIfNeeded();
-    const familyEvidence = await inspectFamily({
-      testFrame,
-      routeFrame,
-      interactiveEntry,
-    });
-
-    const beforeMode = await previewFrame.locator("html").getAttribute("class");
-    await routeFrame.getByRole("button", { name: "Toggle color mode" }).click();
-    await testFrame.waitForTimeout(250);
-    const afterMode = await previewFrame.locator("html").getAttribute("class");
-    const configTheme = await previewFrame.locator("html").getAttribute("data-kb-theme");
-    const firstWidth = await firstEntry.evaluate(
-      (element) => element.getBoundingClientRect().width,
-    );
-    const interactiveWidth = await interactiveEntry.evaluate(
-      (element) => element.getBoundingClientRect().width,
-    );
-    const previewHeight = await firstEntry
-      .locator("[data-chart-preview]")
-      .evaluate((element) => element.getBoundingClientRect().height);
-
-    return {
-      heading: await heading.textContent(),
-      routeCanonicalPath: await routeFrame
-        .locator("main[data-product-surface=charts]")
-        .getAttribute("data-canonical-route"),
-      entryCount: await entries.count(),
-      previewHeight,
-      previewLoading: await firstEntry.locator("iframe").getAttribute("loading"),
-      interactiveIsFullWidth: interactiveWidth > firstWidth * 2,
-      previews,
-      keyboardTooltipText,
-      pointerTooltipText,
-      colorModeSynchronized: beforeMode !== afterMode,
-      configThemeSynchronized: configTheme === (afterMode?.includes("dark") ? "dark" : "light"),
-      rtlHasNoOverflow,
-      consoleErrors,
-      ...familyEvidence,
-    };
-  } finally {
-    page.off("console", recordConsole);
-    page.off("pageerror", recordPageError);
-  }
-}
-
-async function inspectDeferredChartPreviews(context: unknown, fixture: ChartCatalogFixture) {
-  const testFrame = await getBrowserCommandFrame(context);
-  const routeFrame = testFrame.frameLocator(`iframe[title="${fixture.routeTitle}"]`);
-  const entries = routeFrame.locator("[data-chart-entry]");
-  await entries.first().waitFor({ state: "visible" });
-  await testFrame.waitForTimeout(6_500);
-
-  let deferredAlertCount = 0;
-  for (let index = 6; index < (await entries.count()); index += 1) {
-    deferredAlertCount += await entries.nth(index).getByRole("alert").count();
-  }
-
-  return {
-    alertCount: await routeFrame.getByRole("alert").count(),
-    deferredAlertCount,
-    hasNoOverflow: await routeFrame
-      .locator("html")
-      .evaluate((element) => element.scrollWidth <= element.clientWidth),
-  };
-}
-
-async function exerciseChartSourceFailure(context: unknown, fixture: ChartCatalogFixture) {
-  const testFrame = await getBrowserCommandFrame(context);
-  const page = testFrame.page();
-  const routeFrame = testFrame.frameLocator(`iframe[title="${fixture.routeTitle}"]`);
-  const firstEntry = routeFrame.locator(`[data-chart-entry="${fixture.firstSlug}"]`);
-  const sourceChunk = new RegExp(`/assets/${fixture.firstSlug}-[^/]+\\.tsx(?:\\?.*)?$`);
-  const pageErrors: string[] = [];
-  const recordPageError = (error: Error) => pageErrors.push(error.message);
-  page.on("pageerror", recordPageError);
-
-  try {
-    await firstEntry.waitFor({ state: "visible" });
-    await firstEntry.scrollIntoViewIfNeeded();
-    await firstEntry.frameLocator("iframe").locator('[data-slot="chart"]').waitFor({
-      state: "visible",
-    });
-    await page.route(
-      sourceChunk,
-      (route) => route.fulfill({ status: 503, body: "Source temporarily unavailable" }),
-      { times: 1 },
-    );
-
-    await firstEntry.getByRole("button", { name: "View Code" }).click();
-    const alert = routeFrame.getByRole("alert").filter({ hasText: "Source failed to load" });
-    await alert.waitFor({ state: "visible", timeout: 7_000 });
-    const retry = alert.getByRole("button", { name: "Retry source" });
-    const evidence = {
-      alertText: (await alert.textContent()) ?? "",
-      retryVisible: await retry.isVisible(),
-    };
-
-    await page.unroute(sourceChunk);
-    await retry.press("Enter");
-    const source = routeFrame.locator("pre code");
-    await source.waitFor({ state: "visible", timeout: 7_000 });
-    await source.getByText(`export function ${fixture.firstExportName}`, { exact: false }).waitFor({
-      state: "visible",
-    });
-    return { ...evidence, recovered: true, pageErrors };
-  } finally {
-    page.off("pageerror", recordPageError);
-    await page.unroute(sourceChunk);
-  }
-}
-
-async function exerciseChartPreviewFailure(context: unknown, fixture: ChartCatalogFixture) {
-  const testFrame = await getBrowserCommandFrame(context);
-  const page = testFrame.page();
-  const routeFrame = testFrame.frameLocator(`iframe[title="${fixture.routeTitle}"]`);
-  const blockedPreview = `**/preview/charts/${fixture.firstSlug}`;
-  await page.route(blockedPreview, (route) => route.abort());
-  try {
-    await routeFrame.locator("body").evaluate(() => window.location.reload());
-    await testFrame.waitForTimeout(750);
-    const firstEntry = routeFrame.locator(`[data-chart-entry="${fixture.firstSlug}"]`);
-    await firstEntry.waitFor({ state: "visible" });
-    await firstEntry.scrollIntoViewIfNeeded();
-    const alert = firstEntry.getByRole("alert");
-    await alert.waitFor({ state: "visible", timeout: 8_000 });
-    const retry = alert.getByRole("button", { name: /Retry/ });
-    const openPreview = alert.getByRole("link", { name: /Open Preview/ });
-    const evidence = {
-      alertText: (await alert.textContent()) ?? "",
-      retryVisible: await retry.isVisible(),
-      openPreviewVisible: await openPreview.isVisible(),
-    };
-
-    await page.unroute(blockedPreview);
-    await retry.click();
-    const recoveredFrame = routeFrame.frameLocator(
-      `iframe[title="${fixture.recoveredPreviewTitle}"]`,
-    );
-    await recoveredFrame.locator('[data-slot="chart"]').waitFor({
-      state: "visible",
-      timeout: 8_000,
-    });
-    return { ...evidence, recovered: true };
-  } finally {
-    await page.unroute(blockedPreview);
-  }
 }
 
 export default defineConfig({
@@ -519,541 +248,6 @@ export default defineConfig({
                 };
               },
               ...createBrowserCommands,
-              async inspectBuiltChart(context) {
-                const providerContext = context.provider.getCommandsContext(context.sessionId) as {
-                  frame: () => Promise<Frame>;
-                };
-                const testFrame = await providerContext.frame();
-                const routeFrame = testFrame.frameLocator('iframe[title="Built chart route"]');
-                const chart = routeFrame.locator('[data-slot="chart"]');
-
-                await chart.waitFor({ state: "visible" });
-                await chart.locator(".recharts-surface").waitFor({ state: "visible" });
-
-                return {
-                  chartSlot: await chart.getAttribute("data-chart"),
-                  description: await routeFrame
-                    .getByText("Showing total visitors for the last 6 months")
-                    .textContent(),
-                  renderedAreaCount: await chart.locator(".recharts-area-area").count(),
-                };
-              },
-              async inspectAreaChartCatalog(context) {
-                return inspectChartCatalogFamily(
-                  context,
-                  chartCatalogFixtures.area,
-                  async ({ routeFrame, interactiveEntry }) => {
-                    const interactiveFrame = interactiveEntry.frameLocator("iframe");
-                    const rangeTrigger = interactiveFrame.getByRole("button", {
-                      name: "Select a value",
-                    });
-                    await rangeTrigger.click();
-                    await interactiveFrame.getByRole("option", { name: "Last 7 days" }).waitFor({
-                      state: "visible",
-                    });
-                    await rangeTrigger.press("End");
-                    await rangeTrigger.press("Enter");
-                    const interactiveSelection = await rangeTrigger.textContent();
-                    return {
-                      families: await routeFrame
-                        .getByRole("navigation", { name: "Chart families" })
-                        .getByRole("link")
-                        .allTextContents(),
-                      interactiveSelection: interactiveSelection?.trim() ?? null,
-                    };
-                  },
-                );
-              },
-              async inspectRadarChartCatalog(context) {
-                const providerContext = context.provider.getCommandsContext(context.sessionId) as {
-                  frame: () => Promise<Frame>;
-                };
-                const testFrame = await providerContext.frame();
-                const page = testFrame.page();
-                const routeFrame = testFrame.frameLocator(
-                  'iframe[title="Radar Chart Catalog route"]',
-                );
-                const consoleErrors: string[] = [];
-                const recordConsole = (message: { type: () => string; text: () => string }) => {
-                  if (message.type() === "error") consoleErrors.push(message.text());
-                };
-                const recordPageError = (error: Error) => consoleErrors.push(error.message);
-                page.on("console", recordConsole);
-                page.on("pageerror", recordPageError);
-
-                try {
-                  await routeFrame.locator("body").evaluate(() => window.location.reload());
-                  await testFrame.waitForTimeout(500);
-                  const main = routeFrame.locator('[data-canonical-route="/charts/radar"]');
-                  await main.waitFor({ state: "visible" });
-                  const entries = routeFrame.locator("[data-chart-entry]");
-                  const previews: Array<{
-                    chartRole: string | null;
-                    circleGridCount: number;
-                    polygonCount: number;
-                    polarGridCount: number;
-                    radialLineCount: number;
-                    slug: string | null;
-                  }> = [];
-
-                  for (let index = 0; index < (await entries.count()); index += 1) {
-                    const catalogEntry = entries.nth(index);
-                    await catalogEntry.scrollIntoViewIfNeeded();
-                    const preview = catalogEntry.frameLocator("iframe");
-                    const chart = preview.locator('[data-slot="chart"]');
-                    await chart.waitFor({ state: "visible", timeout: 8_000 });
-                    await chart.locator(".recharts-surface").waitFor({ state: "visible" });
-                    previews.push({
-                      slug: await catalogEntry.getAttribute("data-chart-entry"),
-                      polygonCount: await chart.locator(".recharts-radar-polygon").count(),
-                      polarGridCount: await chart.locator(".recharts-polar-grid").count(),
-                      circleGridCount: await chart
-                        .locator(".recharts-polar-grid-concentric-circle")
-                        .count(),
-                      radialLineCount: await chart
-                        .locator(".recharts-polar-grid-angle line")
-                        .count(),
-                      chartRole: await preview.getByRole("application").getAttribute("role"),
-                    });
-                  }
-
-                  const firstEntry = entries.first();
-                  await firstEntry.scrollIntoViewIfNeeded();
-                  const firstPreview = firstEntry.frameLocator("iframe");
-                  const firstChart = firstPreview.locator('[data-slot="chart"]');
-                  const application = firstPreview.getByRole("application");
-                  await application.focus();
-                  await page.keyboard.press("ArrowRight");
-                  await testFrame.waitForTimeout(250);
-                  const keyboardTooltipText = await firstPreview
-                    .locator(".cn-chart-tooltip")
-                    .textContent();
-                  await firstChart
-                    .locator(".recharts-radar-polygon")
-                    .hover({ position: { x: 20, y: 20 } });
-                  await firstPreview.locator(".cn-chart-tooltip").waitFor({ state: "visible" });
-                  const pointerTooltipText = await firstPreview
-                    .locator(".cn-chart-tooltip")
-                    .textContent();
-
-                  const iconsEntry = routeFrame.locator('[data-chart-entry="chart-radar-icons"]');
-                  await iconsEntry.scrollIntoViewIfNeeded();
-                  const iconsPreview = iconsEntry.frameLocator("iframe");
-                  const legend = iconsPreview.locator(".recharts-legend-wrapper");
-                  await legend.waitFor({ state: "visible" });
-                  await testFrame.waitForTimeout(1_800);
-                  const firstCapture = await iconsPreview.locator("main").screenshot();
-                  await testFrame.waitForTimeout(250);
-                  const secondCapture = await iconsPreview.locator("main").screenshot();
-
-                  await iconsEntry.getByRole("button", { name: "View Code" }).click();
-                  const sourceDialog = routeFrame.getByRole("dialog");
-                  await sourceDialog.waitFor({ state: "visible" });
-                  const sourceCode = sourceDialog.locator("pre code");
-                  await sourceCode
-                    .getByText("export function ChartRadarIcons", { exact: false })
-                    .waitFor({
-                      state: "visible",
-                    });
-                  const installCommand = await sourceDialog
-                    .getByText("bunx shadcn@latest add @zaidan/chart-radar-icons", { exact: true })
-                    .textContent();
-                  const sourceLoaded = (await sourceCode.textContent())?.includes(
-                    "export function ChartRadarIcons",
-                  );
-                  await page.keyboard.press("Escape");
-                  await sourceDialog.waitFor({ state: "hidden" });
-
-                  const customLabelPreview = routeFrame
-                    .locator('[data-chart-entry="chart-radar-label-custom"]')
-                    .frameLocator("iframe");
-                  const customLabels = await customLabelPreview
-                    .locator(".recharts-polar-angle-axis-tick")
-                    .allTextContents();
-                  const firstPreviewHeight = await firstEntry
-                    .locator("[data-chart-preview]")
-                    .evaluate((element) => element.getBoundingClientRect().height);
-
-                  return {
-                    canonicalPath: await main.getAttribute("data-canonical-route"),
-                    activeFamily: await routeFrame
-                      .getByRole("navigation", { name: "Chart families" })
-                      .locator('[aria-current="page"]')
-                      .textContent(),
-                    entryCount: await entries.count(),
-                    previewHeight: firstPreviewHeight,
-                    previewLoading: await firstEntry.locator("iframe").getAttribute("loading"),
-                    previews,
-                    keyboardTooltipText,
-                    pointerTooltipText,
-                    iconLegendLabels: await legend.locator(":scope > div > div").allTextContents(),
-                    iconLegendSvgCount: await legend.locator("svg").count(),
-                    sharedActionCount: await entries
-                      .filter({ has: routeFrame.getByRole("button", { name: "Copy Code" }) })
-                      .count(),
-                    installCommand,
-                    sourceLoaded,
-                    customLabels: customLabels.join(" "),
-                    representativeScreenshotBytes: firstCapture.length,
-                    representativeDeterministic: firstCapture.equals(secondCapture),
-                    consoleErrors,
-                  };
-                } finally {
-                  page.off("console", recordConsole);
-                  page.off("pageerror", recordPageError);
-                }
-              },
-              async inspectDeferredAreaPreviews(context) {
-                return inspectDeferredChartPreviews(context, chartCatalogFixtures.area);
-              },
-              async exerciseAreaSourceFailure(context) {
-                return exerciseChartSourceFailure(context, chartCatalogFixtures.area);
-              },
-              async exerciseAreaChartFailure(context) {
-                return exerciseChartPreviewFailure(context, chartCatalogFixtures.area);
-              },
-              async inspectLineChartCatalog(context) {
-                return inspectChartCatalogFamily(
-                  context,
-                  chartCatalogFixtures.line,
-                  async ({ testFrame, routeFrame, interactiveEntry }) => {
-                    const interactiveFrame = interactiveEntry.frameLocator("iframe");
-                    const mobileSeries = interactiveFrame.getByRole("button", { name: /Mobile/ });
-                    await mobileSeries.press("Enter");
-                    await testFrame.waitForTimeout(100);
-                    const interactiveSelection = await interactiveFrame
-                      .getByRole("button", { pressed: true })
-                      .textContent();
-
-                    const representativeFrame = routeFrame
-                      .locator('[data-chart-entry="chart-line-dots-custom"]')
-                      .frameLocator("iframe");
-                    const customDotCount = await representativeFrame
-                      .locator("[data-custom-line-dot]")
-                      .count();
-                    const labelFrame = routeFrame
-                      .locator('[data-chart-entry="chart-line-label"]')
-                      .frameLocator("iframe");
-                    const labelCount = await labelFrame.locator("[data-line-label]").count();
-                    const customLabelFrame = routeFrame
-                      .locator('[data-chart-entry="chart-line-label-custom"]')
-                      .frameLocator("iframe");
-                    const customLabels = await customLabelFrame
-                      .locator("[data-line-label]")
-                      .allTextContents();
-
-                    return {
-                      activeFamily: await routeFrame
-                        .getByRole("navigation", { name: "Chart families" })
-                        .locator('[aria-current="page"]')
-                        .textContent(),
-                      interactiveSelection: interactiveSelection?.trim() ?? null,
-                      customDotCount,
-                      labelCount,
-                      customLabels,
-                    };
-                  },
-                );
-              },
-              async inspectDeferredLinePreviews(context) {
-                return inspectDeferredChartPreviews(context, chartCatalogFixtures.line);
-              },
-              async exerciseLineSourceFailure(context) {
-                return exerciseChartSourceFailure(context, chartCatalogFixtures.line);
-              },
-              async exerciseLineChartFailure(context) {
-                return exerciseChartPreviewFailure(context, chartCatalogFixtures.line);
-              },
-              async exerciseRadarChartFailure(context) {
-                return exerciseChartPreviewFailure(context, chartCatalogFixtures.radar);
-              },
-              async inspectRadarRepresentative(context) {
-                const providerContext = context.provider.getCommandsContext(context.sessionId) as {
-                  frame: () => Promise<Frame>;
-                };
-                const testFrame = await providerContext.frame();
-                const route = testFrame.locator('iframe[title="Radar representative Preview"]');
-                const routeFrame = route.contentFrame();
-                await route.evaluate((element) => {
-                  element.style.border = "0";
-                });
-                const main = routeFrame.locator('[data-preview-slug="chart-radar-icons"]');
-                await main.waitFor({ state: "visible" });
-                await routeFrame.locator(".recharts-surface").waitFor({ state: "visible" });
-                await routeFrame.locator("html").evaluate((element) => {
-                  element.classList.remove("dark");
-                  element.classList.add("light");
-                  element.dataset.kbTheme = "light";
-                });
-                await testFrame.waitForTimeout(1_800);
-                const lightCapture = await route.screenshot();
-                const lightViewport = await route.evaluate((element) => {
-                  const bounds = element.getBoundingClientRect();
-                  return { width: bounds.width, height: bounds.height };
-                });
-
-                await route.evaluate((element) => {
-                  element.style.width = "1440px";
-                  element.style.height = "900px";
-                });
-                await routeFrame.locator("html").evaluate((element) => {
-                  element.classList.remove("light");
-                  element.classList.add("dark");
-                  element.dataset.kbTheme = "dark";
-                });
-                await testFrame.waitForTimeout(250);
-                const darkCapture = await route.screenshot();
-                await testFrame.waitForTimeout(250);
-                const repeatedDarkCapture = await route.screenshot();
-                const darkViewport = await route.evaluate((element) => {
-                  const bounds = element.getBoundingClientRect();
-                  return { width: bounds.width, height: bounds.height };
-                });
-
-                return {
-                  slug: await main.getAttribute("data-preview-slug"),
-                  legendLabels: await routeFrame
-                    .locator(".recharts-legend-wrapper > div > div")
-                    .allTextContents(),
-                  lightViewport,
-                  darkViewport,
-                  lightBytes: lightCapture.length,
-                  darkBytes: darkCapture.length,
-                  darkStable: darkCapture.equals(repeatedDarkCapture),
-                };
-              },
-              async inspectTooltipChartCatalog(context) {
-                const providerContext = context.provider.getCommandsContext(context.sessionId) as {
-                  frame: () => Promise<Frame>;
-                };
-                const testFrame = await providerContext.frame();
-                const page = testFrame.page();
-                const routeFrame = testFrame.frameLocator('iframe[title="Chart Catalog route"]');
-                const consoleErrors: string[] = [];
-                const recordConsole = (message: { type: () => string; text: () => string }) => {
-                  if (message.type() === "error") consoleErrors.push(message.text());
-                };
-                const recordPageError = (error: Error) => consoleErrors.push(error.message);
-                page.on("console", recordConsole);
-                page.on("pageerror", recordPageError);
-
-                try {
-                  await routeFrame.locator("body").evaluate(() => window.location.reload());
-                  await testFrame.waitForTimeout(500);
-
-                  const entries = routeFrame.locator("[data-chart-entry]");
-                  await entries.first().waitFor({ state: "visible" });
-                  const slugs: Array<string | null> = [];
-                  const labels: string[] = [];
-                  const actionCounts: number[] = [];
-                  const renderedBarCounts: number[] = [];
-                  const applicationRoles: Array<string | null> = [];
-                  let defaultTooltipText: string | null = null;
-                  let keyboardTooltipText: string | null = null;
-                  let pointerTooltipText: string | null = null;
-                  let customLabelText: string | null = null;
-                  let labelFormatterText: string | null = null;
-                  let formatterText: string | null = null;
-                  let iconCount = 0;
-                  let advancedText: string | null = null;
-                  let advancedCaptureBytes = 0;
-
-                  for (let index = 0; index < (await entries.count()); index += 1) {
-                    const entry = entries.nth(index);
-                    await entry.scrollIntoViewIfNeeded();
-                    const preview = entry.frameLocator("iframe");
-                    const chart = preview.locator('[data-slot="chart"]');
-                    await chart.waitFor({ state: "visible", timeout: 8_000 });
-                    await chart.locator(".recharts-surface").waitFor({ state: "visible" });
-                    const slug = await entry.getAttribute("data-chart-entry");
-                    slugs.push(slug);
-                    labels.push(
-                      (await entry.getByRole("heading", { level: 3 }).textContent())?.trim() ?? "",
-                    );
-                    actionCounts.push(
-                      await entry.locator("[data-chart-actions]").locator("button, a").count(),
-                    );
-                    renderedBarCounts.push(await chart.locator(".recharts-bar-rectangle").count());
-                    applicationRoles.push(
-                      await preview.getByRole("application").getAttribute("role"),
-                    );
-                    const tooltip = preview.locator(".cn-chart-tooltip");
-                    await tooltip.waitFor({ state: "visible", timeout: 8_000 });
-                    if (slug === "chart-tooltip-default") {
-                      defaultTooltipText = await tooltip.textContent();
-                      const application = preview.getByRole("application");
-                      await application.focus();
-                      await page.keyboard.press("ArrowRight");
-                      await testFrame.waitForTimeout(150);
-                      keyboardTooltipText = await tooltip.textContent();
-                      await chart
-                        .locator(".recharts-surface")
-                        .hover({ position: { x: 220, y: 150 } });
-                      await tooltip.waitFor({ state: "visible" });
-                      pointerTooltipText = await tooltip.textContent();
-                    } else if (slug === "chart-tooltip-label-custom") {
-                      customLabelText = await tooltip.textContent();
-                    } else if (slug === "chart-tooltip-label-formatter") {
-                      labelFormatterText = await tooltip.textContent();
-                    } else if (slug === "chart-tooltip-formatter") {
-                      formatterText = await tooltip.textContent();
-                    } else if (slug === "chart-tooltip-icons") {
-                      iconCount = await tooltip.locator("svg").count();
-                    } else if (slug === "chart-tooltip-advanced") {
-                      advancedText = await tooltip.textContent();
-                      advancedCaptureBytes = (await preview.locator("body").screenshot()).length;
-                    }
-                  }
-
-                  const firstEntry = entries.first();
-                  const activeFamily = await routeFrame
-                    .getByRole("navigation", { name: "Chart families" })
-                    .locator('[aria-current="page"]')
-                    .textContent();
-                  const viewCode = firstEntry.getByRole("button", { name: "View Code" });
-                  await viewCode.focus();
-                  await page.keyboard.press("Enter");
-                  const installCommand = routeFrame.getByText(
-                    "bunx shadcn@latest add @zaidan/chart-tooltip-default",
-                    { exact: true },
-                  );
-                  await installCommand.waitFor({ state: "visible" });
-                  const source = routeFrame.locator("pre code");
-                  await source
-                    .getByText("export function ChartTooltipDefault", { exact: false })
-                    .waitFor({ state: "visible", timeout: 7_000 });
-
-                  const advancedFrameElement = await routeFrame
-                    .locator('[data-chart-entry="chart-tooltip-advanced"] iframe')
-                    .elementHandle();
-                  const advancedFrame = await advancedFrameElement?.contentFrame();
-                  if (!advancedFrame) throw new TypeError("Advanced Tooltip frame is unavailable");
-                  await advancedFrame.addScriptTag({ content: axeCore.source });
-                  const accessibilityViolations = await advancedFrame.evaluate(async () => {
-                    const axe = (
-                      window as typeof window & {
-                        axe: {
-                          run: (
-                            root: Document,
-                            options: { runOnly: { type: "tag"; values: string[] } },
-                          ) => Promise<{
-                            violations: {
-                              id: string;
-                              impact: string | null;
-                              nodes: { target: string[] }[];
-                            }[];
-                          }>;
-                        };
-                      }
-                    ).axe;
-                    const results = await axe.run(document, {
-                      runOnly: {
-                        type: "tag",
-                        values: ["wcag2a", "wcag2aa", "wcag21a", "wcag21aa", "wcag22aa"],
-                      },
-                    });
-                    return results.violations.map((violation) => ({
-                      id: violation.id,
-                      impact: violation.impact,
-                      targets: violation.nodes.map(({ target }) => target),
-                    }));
-                  });
-
-                  return {
-                    canonicalRoute: await routeFrame
-                      .locator("main")
-                      .getAttribute("data-canonical-route"),
-                    activeFamily,
-                    labels,
-                    slugs,
-                    actionCounts,
-                    previewHeight: await firstEntry
-                      .locator("[data-chart-preview]")
-                      .evaluate((element) => element.getBoundingClientRect().height),
-                    lazyPreviewCount: await entries.locator('iframe[loading="lazy"]').count(),
-                    renderedBarCounts,
-                    applicationRoles,
-                    defaultTooltipText,
-                    keyboardTooltipText,
-                    pointerTooltipText,
-                    customLabelText,
-                    labelFormatterText,
-                    formatterText,
-                    iconCount,
-                    advancedText,
-                    advancedCaptureBytes,
-                    accessibilityViolations,
-                    installCommand: await installCommand.textContent(),
-                    sourceContainsExport: ((await source.textContent()) ?? "").includes(
-                      "export function ChartTooltipDefault",
-                    ),
-                    consoleErrors,
-                  };
-                } finally {
-                  page.off("console", recordConsole);
-                  page.off("pageerror", recordPageError);
-                }
-              },
-              async exerciseTooltipPreviewRecovery(context) {
-                const providerContext = context.provider.getCommandsContext(context.sessionId) as {
-                  frame: () => Promise<Frame>;
-                };
-                const testFrame = await providerContext.frame();
-                const page = testFrame.page();
-                const routeFrame = testFrame.frameLocator('iframe[title="Chart Catalog route"]');
-                const blockedPreview = "**/preview/charts/chart-tooltip-default";
-                const pageErrors: string[] = [];
-                const recordPageError = (error: Error) => pageErrors.push(error.message);
-                page.on("pageerror", recordPageError);
-
-                try {
-                  const firstEntry = routeFrame.locator(
-                    '[data-chart-entry="chart-tooltip-default"]',
-                  );
-                  await firstEntry.waitFor({ state: "visible" });
-                  await firstEntry.scrollIntoViewIfNeeded();
-                  await firstEntry
-                    .frameLocator("iframe")
-                    .locator('[data-slot="chart"]')
-                    .waitFor({ state: "visible" });
-                  await firstEntry.locator("iframe").dispatchEvent("error");
-
-                  const alert = firstEntry.getByRole("alert");
-                  await alert.waitFor({ state: "visible", timeout: 8_000 });
-                  const retry = alert.getByRole("button", { name: /Retry/ });
-                  const openPreview = alert.getByRole("link", { name: /Open Preview/ });
-                  const evidence = {
-                    alertText: (await alert.textContent()) ?? "",
-                    retryVisible: await retry.isVisible(),
-                    openPreviewVisible: await openPreview.isVisible(),
-                  };
-
-                  await page.route(
-                    blockedPreview,
-                    async (route) => {
-                      await testFrame.waitForTimeout(500);
-                      await route.abort();
-                    },
-                    { times: 1 },
-                  );
-                  await retry.click();
-                  const loading = firstEntry.getByRole("status");
-                  await loading.waitFor({ state: "visible" });
-                  const loadingText = (await loading.textContent()) ?? "";
-                  await alert.waitFor({ state: "visible", timeout: 8_000 });
-
-                  await page.unroute(blockedPreview);
-                  await retry.click();
-                  await firstEntry
-                    .frameLocator('iframe[title="Tooltip — Default Preview"]')
-                    .locator('[data-slot="chart"]')
-                    .waitFor({ state: "visible", timeout: 8_000 });
-                  return { ...evidence, loadingText, recovered: true, pageErrors };
-                } finally {
-                  page.off("pageerror", recordPageError);
-                  await page.unroute(blockedPreview);
-                }
-              },
               async inspectDesktopProductHeader(context) {
                 const { header } = await getProductRouteTestFrame(context);
